requirements.txt
opencv-python>=4.8.0
mysql-connector-python>=8.0.33
Pillow>=10.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.1
